requirements.txt
asgiref==3.8.1
Django==5.1.2
django-allauth==65.1.0
django-environ==0.11.2
Faker==30.8.1
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
